requirements.txt
Django~=5.2
pillow==12.2.*
django-admin-sortable2==2.3.*
django-tinymce==5.0.*
environs==15.0.*
requests==2.33.*